+++ libs/oidc/src/oidc.constants.ts
@@ -1,6 +1,8 @@
 export const OIDC_MODULE_OPTIONS = 'OidcModuleOptions';
 export const TOKEN_STORE = 'TokenStore';
 export const defaultModuleOptions = {
-  clockTolerance: 10,
+  authParams: {
+    clockTolerance: 10,
+  },
   userInfoMethod: 'ffdc',
 };