+++ libs/oidc/src/utils/merge-defaults.util.ts
@@ -2,8 +2,13 @@ import { OidcModuleOptions } from '../interfaces';
 import { defaultModuleOptions } from '../oidc.constants';
 
 export function mergeDefaults(options: OidcModuleOptions) {
-  return {
+  const newOptions = {
     ...defaultModuleOptions,
     ...options,
   } as OidcModuleOptions;
+  newOptions.authParams = {
+    ...defaultModuleOptions.authParams,
+    ...options.authParams,
+  };
+  return newOptions;
 }