+++ libs/proxy/src/utils/string.utils.ts
@@ -1,0 +1,3 @@
+export function endsWith(str, suffix) {
+  return str.indexOf(suffix, str.length - suffix.length) !== -1;
+}