+++ src/app.module.ts
@@ -1,7 +1,9 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { CatsModule } from './cats/cats.module';
-import { OidcModule } from '@uxd-finastra/oidc';
+import { OidcModule } from '@ffdc/nestjs-oidc';
+import { ProxyModule } from '@ffdc/nestjs-proxy';
+import { ProxyConfigService } from './proxy-config.service';
 
 @Module({
   imports: [
@@ -12,16 +14,23 @@ import { OidcModule } from '@uxd-finastra/oidc';
       imports: [ConfigModule.forRoot()],
       useFactory: async (configService: ConfigService) => ({
         issuer: configService.get('OIDC_ISSUER'),
-        clientId: configService.get('OIDC_CLIENT_ID'),
-        clientSecret: configService.get('OIDC_CLIENT_SECRET'),
-        scopes: configService.get('OIDC_SCOPES'),
-        redirectUriLogin: configService.get('OIDC_LOGIN_REDIRECT_URI'),
-        redirectUriLogout: configService.get('OIDC_LOGOUT_REDIRECT_URI'),
-        resource: configService.get('OIDC_RESOURCE'),
+        clientMetadata: {
+          client_id: configService.get('OIDC_CLIENT_ID'),
+          client_secret: configService.get('OIDC_CLIENT_SECRET'),
+        },
+        authParams: {
+          scopes: configService.get('OIDC_SCOPES'),
+          resource: configService.get('OIDC_RESOURCE'),
+        },
+        origin: configService.get('OIDC_ORIGIN'),
       }),
       inject: [ConfigService],
     }),
     CatsModule,
+    ProxyModule.forRootAsync(ProxyModule, {
+      useClass: ProxyConfigService,
+      imports: [ConfigModule.forRoot()],
+    }),
   ],
   controllers: [],
   providers: [],